remove tempdb in between file imports
